refactor : 수정사항 반영 (#166)
* 중간 저장

* refactor : 수정사항 반영

* fix : 로직 오류 수정

* refactor: N+1 및 Race Condition 방지

* fix : 피드 조회 수정

* feat : 키워드 api 구현

+++ src/main/java/com/example/cp_main_be/global/dto/FeedItemResponse.java
@@ -9,9 +9,9 @@ public interface FeedItemResponse {
 
   AuthorResponse getAuthor();
 
-  int getLikeCount();
+  long getLikeCount();
 
-  int getCommentCount();
+  long getCommentCount();
 
   LocalDateTime getCreatedAt();
 }
